Refactoring of the blending operations
Split blend.c in several files (separated by "color space") to make the
code manageable.
