+++ src/CMakeLists.txt
@@ -101,6 +101,9 @@ FILE(GLOB SOURCE_FILES
   "develop/pixelpipe.c"
   "develop/blend.c"
   "develop/blend_gui.c"
+  "develop/blends/blendif_lab.c"
+  "develop/blends/blendif_raw.c"
+  "develop/blends/blendif_rgb_hsl.c"
   "develop/tiling.c"
   "common/dwt.c"
   "common/heal.c"
